extract board-context module
