+++ apps/server/src/modules/board-context/board-context-api-helper.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { RoomModule } from '../room';
 import { BoardContextApiHelperService } from './board-context-api-helper.service';
-import { BoardModule } from './board.module';
+import { BoardModule } from '../board/board.module';
 import { LearnroomModule } from '../learnroom';
 
 @Module({
