+++ apps/server/src/modules/board-context/index.ts
@@ -1,0 +1,2 @@
+export { BoardContextApiHelperModule } from './board-context-api-helper.module';
+export { BoardContextApiHelperService } from './board-context-api-helper.service';
