+++ apps/server/src/modules/tool/context-external-tool/uc/context-external-tool.uc.ts
@@ -10,7 +10,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import { User } from '@shared/domain/entity';
 import { Permission } from '@shared/domain/interface';
 import { EntityId } from '@shared/domain/types';
-import { BoardContextApiHelperService } from '@modules/board/board-context-api-helper.service';
+import { BoardContextApiHelperService } from '@modules/board-context';
 import { Authorization } from 'oauth-1.0a';
 import { ToolConfigType, ToolContextType } from '../../common/enum';
 import { Lti11EncryptionService } from '../../common/service';
