+++ apps/server/src/modules/tool/external-tool/uc/external-tool-configuration.uc.ts
@@ -5,7 +5,7 @@ import { Page } from '@shared/domain/domainobject/page';
 import { User } from '@shared/domain/entity';
 import { Permission } from '@shared/domain/interface';
 import { EntityId } from '@shared/domain/types';
-import { BoardContextApiHelperService } from '@modules/board/board-context-api-helper.service';
+import { BoardContextApiHelperService } from '@modules/board-context';
 import { School, SchoolService } from '@modules/school';
 import { CustomParameterScope, ToolContextType } from '../../common/enum';
 import { ToolPermissionHelper } from '../../common/uc/tool-permission-helper';
