+++ apps/server/src/modules/tool/tool-api.module.ts
@@ -1,7 +1,7 @@
 import { EncryptionModule } from '@infra/encryption';
 import { AuthorizationModule } from '@modules/authorization';
 import { BoardModule } from '@modules/board';
-import { BoardContextApiHelperModule } from '@modules/board/board-context-api-helper.module';
+import { BoardContextApiHelperModule } from '@modules/board-context';
 import { LegacySchoolModule } from '@modules/legacy-school';
 import { SchoolModule } from '@modules/school';
 import { UserModule } from '@modules/user';
